requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dash==3.0.4
diffusers==0.33.1
distro==1.9.0
dlib==20.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
firebase-admin==6.9.0
Flask==3.0.3
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
mysql-connector-python==9.3.0
narwhals==1.41.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
openai==1.82.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
plotly==6.1.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.3
PyMySQL==1.1.1
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
shortuuid==1.0.13
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.0.6
zipp==3.22.0
zstandard==0.23.0
boto3==1.34.82
botocore==1.34.82
Brotli==1.1.0
cffi==1.16.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
jmespath==1.0.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
win_inet_pton==1.1.0
